rename kobuki_control to turtlebot_control
